use a custom controller instead of a override
