+++ spec/cells/decidim/reporting_proposals/admin/edit_note_modal_cell_spec.rb
@@ -2,41 +2,42 @@
 
 require "spec_helper"
 
-describe Decidim::ReportingProposals::EditNoteModalCell, type: :cell do
-  controller Decidim::Proposals::Admin::ProposalsController
+module Decidim::ReportingProposals
+  describe EditNoteModalCell, type: :cell do
+    controller Decidim::Proposals::Admin::ProposalsController
 
-  subject { cell(described_class, model, options).call }
+    subject { cell(described_class, model, options).call }
 
-  let(:proposal_note) { create(:proposal_note) }
-  let(:model) { proposal_note }
-  let(:options) { { modal_id: "editNoteModal" } }
+    let(:proposal_note) { create(:proposal_note) }
+    let(:model) { proposal_note }
+    let(:options) { { modal_id: "editNoteModal" } }
 
+    context "when rendering the note form" do
+      it "renders the note body" do
+        expect(subject).to have_content(proposal_note.body)
+      end
 
-  context "when rendering the note form" do
-    it "renders the note body" do
-      expect(subject).to have_content(proposal_note.body)
+      it "renders the note form" do
+        expect(subject).to have_css("form#edit_proposal_note_#{proposal_note.id}")
+      end
+
+      it "has the modal id set to the specified value or default" do
+        expect(subject).to have_selector("##{options[:modal_id]}")
+      end
     end
 
-    it "renders the note form" do
-      expect(subject).to have_css("form#edit_proposal_note_#{proposal_note.id}")
-    end
+    context "when the proposal note exists" do
+      it "renders the view" do
+        expect(subject).to have_selector("#editNoteModal")
+      end
 
-    it "has the modal id set to the specified value or default" do
-      expect(subject).to have_selector("##{options[:modal_id]}")
-    end
-  end
+      it "assigns the correct proposal to the view" do
+        expect(cell(described_class, model, options).proposal).to eq(Decidim::Proposals::Proposal.find(proposal_note.decidim_proposal_id))
+      end
 
-  context "when the proposal note exists" do
-    it "renders the view" do
-      expect(subject).to have_selector("#editNoteModal")
-    end
-
-    it "assigns the correct proposal to the view" do
-      expect(cell(described_class, model, options).proposal).to eq(Decidim::Proposals::Proposal.find(proposal_note.decidim_proposal_id))
-    end
-
-    it "assigns the correct proposal note body to the view" do
-      expect(cell(described_class, model, options).note_body).to eq(proposal_note.body)
+      it "assigns the correct proposal note body to the view" do
+        expect(cell(described_class, model, options).note_body).to eq(proposal_note.body)
+      end
     end
   end
 end
