+++ spec/lib/overrides_spec.rb
@@ -51,7 +51,6 @@ checksums = [
       "/app/controllers/decidim/proposals/admin/proposal_answers_controller.rb" => "51f25d8b36bca87fe5cb2637149b45fb",
       "/app/views/decidim/proposals/proposals/show.html.erb" => "f27bbec257eb6da28dbdd07ac0a224a5",
       "/app/views/decidim/proposals/proposals/_proposal_similar.html.erb" => "c530c9ee044c1ca3fb9a92c4abcfe3b2",
-      "/app/controllers/decidim/proposals/admin/proposal_notes_controller.rb" => "50e90b25d805770564170dd4b7ea0478"
       "/app/views/decidim/proposals/proposals/_wizard_header.html.erb" => "ff854b9f14b712aa7e1972ca23a54e25"
     }
   }
